:truck: Move module filestorage.adapters.{ => observers.}cookiecutter
