+++ src/cutty/filestorage/adapters/observers/cookiecutter.py
