+++ src/cutty/services/create.py
@@ -8,10 +8,10 @@ from typing import Optional
 
 import appdirs
 
-from cutty.filestorage.adapters.cookiecutter import CookiecutterHooksObserver
 from cutty.filestorage.adapters.disk import DiskFileStorage
 from cutty.filestorage.adapters.disk import FileExistsPolicy
 from cutty.filestorage.adapters.git import GitRepositoryObserver
+from cutty.filestorage.adapters.observers.cookiecutter import CookiecutterHooksObserver
 from cutty.filestorage.domain.files import File
 from cutty.filestorage.domain.observers import FileStorageObserver
 from cutty.filestorage.domain.observers import observe
